Decouple assessor and candidate allocation services
