+++ app/controllers/EventsController.scala
@@ -25,8 +25,8 @@ import model.persisted.eventschedules.EventType.EventType
 import play.api.libs.json.{ JsValue, Json }
 import play.api.mvc.{ Action, AnyContent }
 import repositories.application.GeneralApplicationRepository
-import repositories.events.{ EventsMongoRepository, LocationsWithVenuesInMemoryRepository, LocationsWithVenuesRepository, UnknownVenueException }
-import services.allocation.AssessorAllocationService
+import repositories.events.{ LocationsWithVenuesInMemoryRepository, LocationsWithVenuesRepository, UnknownVenueException }
+import services.allocation.{ AssessorAllocationService, CandidateAllocationService }
 import services.events.EventsService
 import uk.gov.hmrc.play.microservice.controller.BaseController
 
@@ -36,17 +36,16 @@ object EventsController extends EventsController {
   val eventsService: EventsService = EventsService
   val locationsAndVenuesRepository: LocationsWithVenuesRepository = LocationsWithVenuesInMemoryRepository
   val assessorAllocationService: AssessorAllocationService = AssessorAllocationService
+  val candidateAllocationService: CandidateAllocationService = CandidateAllocationService
   val applicationRepository: GeneralApplicationRepository = repositories.applicationRepository
 }
 
 trait EventsController extends BaseController {
   def eventsService: EventsService
-
   def locationsAndVenuesRepository: LocationsWithVenuesRepository
-
   def assessorAllocationService: AssessorAllocationService
-
   def applicationRepository: GeneralApplicationRepository
+  def candidateAllocationService: CandidateAllocationService
 
   def saveAssessmentEvents(): Action[AnyContent] = Action.async { implicit request =>
     eventsService.saveAssessmentEvents().map(_ => Created("Events saved"))
@@ -107,7 +106,7 @@ trait EventsController extends BaseController {
 
   def getEventsWithAllocationsSummary(venueName: String, eventType: EventType): Action[AnyContent] = Action.async { implicit request =>
     locationsAndVenuesRepository.venue(venueName).flatMap { venue =>
-      assessorAllocationService.getEventsWithAllocationsSummary(venue, eventType).map { eventsWithAllocations =>
+      eventsService.getEventsWithAllocationsSummary(venue, eventType).map { eventsWithAllocations =>
         Ok(Json.toJson(eventsWithAllocations))
       }
     }
@@ -115,7 +114,7 @@ trait EventsController extends BaseController {
   def allocateCandidates(eventId: String, sessionId: String): Action[JsValue] = Action.async(parse.json) { implicit request =>
     withJsonBody[exchange.CandidateAllocations] { candidateAllocations =>
       val newAllocations = command.CandidateAllocations.fromExchange(eventId, sessionId, candidateAllocations)
-      assessorAllocationService.allocateCandidates(newAllocations).map {
+      candidateAllocationService.allocateCandidates(newAllocations).map {
         _ => Ok
       }.recover {
         case e: OptimisticLockException => Conflict(e.getMessage)
@@ -124,7 +123,7 @@ trait EventsController extends BaseController {
   }
 
   def getCandidateAllocations(eventId: String, sessionId: String): Action[AnyContent] = Action.async { implicit request =>
-    assessorAllocationService.getCandidateAllocations(eventId, sessionId).map { allocations =>
+    candidateAllocationService.getCandidateAllocations(eventId, sessionId).map { allocations =>
       if (allocations.allocations.isEmpty) {
         NotFound
       } else {
@@ -136,7 +135,7 @@ trait EventsController extends BaseController {
   def removeCandidateAllocations(eventId: String, sessionId: String): Action[JsValue] = Action.async(parse.json) { implicit request =>
     withJsonBody[exchange.CandidateAllocations] { candidateAllocs =>
       val allocations = CandidateAllocation.fromExchange(candidateAllocs, eventId, sessionId).toList
-      assessorAllocationService.unAllocateCandidates(allocations).map( _ => Ok)
+      candidateAllocationService.unAllocateCandidates(allocations).map( _ => Ok)
     }
   }
 }
