+++ app/services/allocation/AssessorAllocationService.scala
@@ -16,22 +16,13 @@
 
 package services.allocation
 
-import common.FutureEx
 import connectors.{ AuthProviderClient, CSREmailClient, EmailClient, ExchangeObjects }
 import model.Exceptions.OptimisticLockException
 import model._
-import model.command.{ AssessorAllocation, AssessorAllocations, CandidateAllocation }
-import model.exchange.{ CandidateAllocationPerSession, EventAssessorAllocationsSummaryPerSkill, EventWithAllocationsSummary }
-import model.persisted.eventschedules.EventType.EventType
-import model.persisted.eventschedules.{ Event, Venue }
-import model.stc.EmailEvents.{ CandidateAllocationConfirmationRequest, CandidateAllocationConfirmed }
-import model.stc.StcEventTypes.StcEvents
-import play.api.Logger
-import play.api.mvc.RequestHeader
+import model.command.{ AssessorAllocation, AssessorAllocations }
+import model.persisted.eventschedules.Event
 import repositories.application.GeneralApplicationRepository
-import repositories.contactdetails.ContactDetailsMongoRepository
-import repositories.personaldetails.PersonalDetailsMongoRepository
-import repositories.{ AssessorAllocationMongoRepository, CandidateAllocationMongoRepository }
+import repositories.{ AssessorAllocationMongoRepository }
 import services.allocation.AssessorAllocationService.CouldNotFindAssessorContactDetails
 import services.events.EventsService
 import services.stc.{ EventSink, StcEventService }
@@ -43,7 +34,6 @@ import scala.concurrent.Future
 object AssessorAllocationService extends AssessorAllocationService {
 
   val assessorAllocationRepo: AssessorAllocationMongoRepository = repositories.assessorAllocationRepository
-  val candidateAllocationRepo: CandidateAllocationMongoRepository = repositories.candidateAllocationRepository
   val applicationRepo: GeneralApplicationRepository = repositories.applicationRepository
 
   val eventsService = EventsService
@@ -58,7 +48,6 @@ object AssessorAllocationService extends AssessorAllocationService {
 trait AssessorAllocationService extends EventSink {
 
   def assessorAllocationRepo: AssessorAllocationMongoRepository
-  def candidateAllocationRepo: CandidateAllocationMongoRepository
   def applicationRepo: GeneralApplicationRepository
 
   def eventsService: EventsService
@@ -68,21 +57,6 @@ trait AssessorAllocationService extends EventSink {
 
   def getAllocations(eventId: String): Future[exchange.AssessorAllocations] = {
     assessorAllocationRepo.allocationsForEvent(eventId).map { a => exchange.AssessorAllocations.apply(a) }
-  }
-
-  def getCandidateAllocations(eventId: String, sessionId: String): Future[exchange.CandidateAllocations] = {
-    candidateAllocationRepo.allocationsForSession(eventId, sessionId).map { a => exchange.CandidateAllocations.apply(a) }
-  }
-
-  def unAllocateCandidates(allocations: List[model.persisted.CandidateAllocation]): Future[SerialUpdateResult[persisted.CandidateAllocation]] = {
-    // For each allocation, reset the progress status and delete the corresponding allocation
-    val res = FutureEx.traverseSerial(allocations) { allocation =>
-      val fut = candidateAllocationRepo.removeCandidateAllocation(allocation).flatMap { _ =>
-        applicationRepo.resetApplicationAllocationStatus(allocation.id)
-      }
-      FutureEx.futureToEither(allocation, fut)
-    }
-    res.map(SerialUpdateResult.fromEither)
   }
 
   def allocate(newAllocations: command.AssessorAllocations)(implicit hc: HeaderCarrier): Future[Unit] = {
@@ -95,9 +69,6 @@ trait AssessorAllocationService extends EventSink {
       case existingAllocations => updateExistingAllocations(existingAllocations, newAllocations).map(_ => ())
     }
   }
-
-  private val dateFormat = "dd MMMM YYYY"
-  private val timeFormat = "HH:mma"
 
   private def notifyNewlyAllocatedAssessors(allocations: command.AssessorAllocations)(implicit hc: HeaderCarrier): Future[Unit] = {
     getContactDetails(allocations).map {
@@ -163,56 +134,6 @@ trait AssessorAllocationService extends EventSink {
     }
   }
 
-  def allocateCandidates(newAllocations: command.CandidateAllocations)(implicit hc: HeaderCarrier, rh: RequestHeader): Future[Unit] = {
-
-    eventsService.getEvent(newAllocations.eventId).flatMap { event =>
-      val eventDate = event.date.toString(dateFormat)
-      val eventTime = event.startTime.toString(timeFormat)
-      val deadlineDateTime = event.date.minusDays(10).toString(dateFormat)
-
-      getCandidateAllocations(newAllocations.eventId, newAllocations.sessionId).flatMap { existingAllocation =>
-        existingAllocation.allocations match {
-          case Nil =>
-            candidateAllocationRepo.save(persisted.CandidateAllocation.fromCommand(newAllocations)).flatMap {
-              _ => Future.sequence(newAllocations.allocations.map(sendCandidateEmail(_, eventDate, eventTime, deadlineDateTime)))
-            }.map(_ => ())
-          case _ =>
-            val existingIds = existingAllocation.allocations.map(_.id)
-            updateExistingAllocations(existingAllocation, newAllocations).flatMap { _ =>
-              Future.sequence(
-                newAllocations.allocations
-                  .filter(alloc => !existingIds.contains(alloc.id))
-                  .map(sendCandidateEmail(_, eventDate, eventTime, deadlineDateTime))
-              )
-            }.map(_ => ())
-        }
-      }
-    }
-  }
-
-  private def sendCandidateEmail(candidateAllocation: CandidateAllocation,
-                                 eventDate: String,
-                                 eventTime: String,
-                                 deadlineDateTime: String)(implicit hc: HeaderCarrier, rh: RequestHeader): Future[Unit] = {
-    applicationRepo.find(candidateAllocation.id).flatMap {
-      case Some(candidate) =>
-        eventSink {
-          val res = authProviderClient.findByUserIds(Seq(candidate.userId)).map { candidates =>
-            candidates.map { candidate =>
-              candidateAllocation.status match {
-                case AllocationStatuses.UNCONFIRMED =>
-                  CandidateAllocationConfirmationRequest(candidate.email, candidate.name, eventDate, eventTime, deadlineDateTime)
-                case AllocationStatuses.CONFIRMED =>
-                  CandidateAllocationConfirmed(candidate.email, candidate.name, eventDate, eventTime)
-              }
-            }
-          } recover { case ex => throw new RuntimeException(s"Was not able to retrieve user details for candidate ${candidate.userId}", ex) }
-          res.asInstanceOf[Future[StcEvents]]
-        }
-      case None => throw new RuntimeException(s"Can not find user application: ${candidateAllocation.id}")
-    }
-  }
-
   private def getAllocationDifferences(existingAllocationsUnsanitised: Seq[persisted.AssessorAllocation],
     newAllocationsUnsanitised: Seq[persisted.AssessorAllocation]) = {
 
@@ -259,52 +180,6 @@ trait AssessorAllocationService extends EventSink {
         _ <- notifyAllocationChangedAssessors(AssessorAllocations(newAllocations.eventId, changedUsers))
         _ <- notifyAllocationUnallocatedAssessors(AssessorAllocations(newAllocations.eventId, removedUsers))
       } yield ()
-    } else {
-      throw OptimisticLockException(s"Stored allocations for event ${newAllocations.eventId} have been updated since reading")
-    }
-  }
-
-  def getEventsWithAllocationsSummary(venue: Venue, eventType: EventType): Future[List[EventWithAllocationsSummary]] = {
-    eventsService.getEvents(eventType, venue).flatMap { events =>
-      val res = events.map { event =>
-        getAllocations(event.id).flatMap { allocations =>
-          val allocationsGroupedBySkill = allocations.allocations.groupBy(_.allocatedAs)
-          val allocationsGroupedBySkillWithSummary = allocationsGroupedBySkill.map { allocationGroupedBySkill =>
-            val assessorAllocation = allocationGroupedBySkill._2
-            val skill = allocationGroupedBySkill._1.name
-            val allocated = assessorAllocation.length
-            val confirmed = assessorAllocation.count(_.status == AllocationStatuses.CONFIRMED)
-            EventAssessorAllocationsSummaryPerSkill(skill, allocated, confirmed)
-          }.toList
-          val candidateAllocBySession = event.sessions.sortBy(_.startTime.getMillisOfDay).map { session =>
-            getCandidateAllocations(event.id, session.id).map { candidateAllocations =>
-              CandidateAllocationPerSession(UniqueIdentifier(session.id),
-                candidateAllocations.allocations.count(_.status == AllocationStatuses.CONFIRMED))
-            }
-          }
-          Future.sequence(candidateAllocBySession).map { cs =>
-            EventWithAllocationsSummary(event.date, event, cs, allocationsGroupedBySkillWithSummary)
-          }
-        }
-      }
-      Future.sequence(res)
-    }
-  }
-
-  private def updateExistingAllocations(existingAllocations: exchange.CandidateAllocations,
-                                        newAllocations: command.CandidateAllocations): Future[Unit] = {
-
-    if (existingAllocations.version.forall(_ == newAllocations.version)) {
-      // no prior update since reading so do update
-      // check what's been updated here so we can send email notifications
-
-      // Convert the existing exchange allocations to persisted objects so we can delete what is currently in the db
-      val toDelete = persisted.CandidateAllocation.fromExchange(existingAllocations, newAllocations.eventId, newAllocations.sessionId)
-
-      val toPersist = persisted.CandidateAllocation.fromCommand(newAllocations)
-      candidateAllocationRepo.delete(toDelete).flatMap { _ =>
-        candidateAllocationRepo.save(toPersist).map(_ => ())
-      }
     } else {
       throw OptimisticLockException(s"Stored allocations for event ${newAllocations.eventId} have been updated since reading")
     }
